finished

+++ controllers/blogs.js
@@ -1,5 +1,6 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
 
 const jwt = require("jsonwebtoken");
 
@@ -12,12 +13,19 @@ const getTokenFrom = (request) => {
 };
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
 
 blogsRouter.post("/", async (request, response) => {
   const body = request.body;
+
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  const user = await User.findById(decodedToken.id);
 
   if (body.title === undefined || body.url === undefined) {
     return response.status(400).end();
@@ -31,6 +39,9 @@ blogsRouter.post("/", async (request, response) => {
   });
 
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
   response.status(201).json(savedBlog);
 });
 
@@ -44,7 +55,28 @@ blogsRouter.get("/:id", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  const user = await User.findById(decodedToken.id);
+
+  const blogToDelete = await Blog.findById(request.params.id);
+
+  if (blogToDelete.author !== user.username) {
+    return response
+      .status(401)
+      .json({ error: "you can delete only your own blogs" });
+  }
+
+  user.blogs = user.blogs.concat();
+  user.blogs = user.blogs.filter((e) => e !== request.params.id);
+  await user.save();
+
   await Blog.findByIdAndRemove(request.params.id);
+
   response.status(204).end();
 });
